requirements.txt
django
djangorestframework-simplejwt
PyJWT==2.1.0